added http command
http command work, using Fetch API and Promise, both polyfilled. The Problem is with body data. For now data are passed as property:value in the command, howeaver if you have to pass string with space it will fail, the parse method in the Interpreter class must be changed to accomodate this instead or just splitting the string by spaces